add zod validation; wrap editor with AnimatePresence

+++ lib/data/queries.ts
@@ -1,4 +1,4 @@
-// TODO: zod validation; error handling
+import { z } from 'zod'
 import { createSupabaseBrowserClient } from '../supabase/client'
 
 export const fetchQueries = async () => {
@@ -7,17 +7,38 @@ export const fetchQueries = async () => {
 }
 
 export const createQuery = async (query: { name: string }) => {
+  const result = z
+    .object({
+      name: z.string().trim().min(1),
+    })
+    .safeParse({ query: query.name })
+
+  if (!result.success) {
+    throw new Error('Invalid query type')
+  }
+
   const supabase = createSupabaseBrowserClient()
   return supabase
     .from('queries')
     .insert({
-      name: query.name,
+      name: result.data.name,
     })
     .select()
     .single()
 }
 
 export const updateQuery = async (query: { id: string; name: string }) => {
+  const result = z
+    .object({
+      id: z.string().uuid(),
+      name: z.string().trim().min(1),
+    })
+    .safeParse({ id: query.id, name: query.name })
+
+  if (!result.success) {
+    throw new Error('Invalid parameters')
+  }
+
   const supabase = createSupabaseBrowserClient()
   return supabase
     .from('queries')
@@ -28,6 +49,12 @@ export const updateQuery = async (query: { id: string; name: string }) => {
 }
 
 export const deleteQuery = async (id: string) => {
+  const result = z.string().uuid().safeParse({ id })
+
+  if (!result.success) {
+    throw new Error('Invalid query id')
+  }
+
   const supabase = createSupabaseBrowserClient()
   return supabase.from('queries').delete().eq('id', id)
 }
